+++ hubdle-project-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/coverage/HubdleConfigCoverageExtension.kt
@@ -15,15 +15,12 @@ import com.javiersc.hubdle.project.extensions.dependencies._internal.constants.O
 import javax.inject.Inject
 import kotlinx.kover.api.CoverageEngineVariant
 import kotlinx.kover.api.IntellijEngine
-import kotlinx.kover.api.KoverMergedConfig
 import kotlinx.kover.api.KoverProjectConfig
 import org.gradle.api.Action
 import org.gradle.api.Project
 import org.gradle.api.Task
 import org.gradle.api.provider.Property
 import org.gradle.kotlin.dsl.the
-import org.gradle.language.base.plugins.LifecycleBasePlugin.BUILD_TASK_NAME
-import org.gradle.language.base.plugins.LifecycleBasePlugin.CHECK_TASK_NAME
 
 @HubdleDslMarker
 public open class HubdleConfigCoverageExtension
@@ -61,37 +58,23 @@ constructor(
         )
 
         configurable {
+            // TODO: Remove this check and `allprojects` when updating to Kover 0.7.0
             check(isRootProject) {
                 "Hubdle `coverage()` must be only configured in the root project"
             }
             allprojects { allproject ->
                 val hubdleKoverEngine = this@HubdleConfigCoverageExtension.engine
                 allproject.the<KoverProjectConfig>().engine.set(hubdleKoverEngine)
-                allproject.the<KoverMergedConfig>().enable()
-            }
-            afterEvaluate {
-                val shouldMergeCodeCoverageReports =
-                    it.gradle.startParameter.taskNames.any { taskName ->
-                        taskName in listOf(ALL_TEST_TASK_NAME, BUILD_TASK_NAME, CHECK_TASK_NAME)
-                    } &&
-                        it.rootProject.tasks.names.contains(KOVER_MERGED_REPORT_TASK_NAME) &&
-                        it.rootProject.tasks.names.contains(ALL_TEST_TASK_NAME)
 
-                if (shouldMergeCodeCoverageReports) {
-                    val koverMergedReportTask =
-                        it.rootProject.tasks.namedLazily<Task>(KOVER_MERGED_REPORT_TASK_NAME)
+                val koverReportTask = allproject.tasks.namedLazily<Task>("koverReport")
 
-                    it.rootProject.tasks.namedLazily<Task>(ALL_TEST_TASK_NAME).configureEach { task
-                        ->
-                        task.dependsOn(koverMergedReportTask)
-                    }
+                allproject.tasks.namedLazily<Task>(ALL_TEST_TASK_NAME).configureEach { task ->
+                    task.dependsOn(koverReportTask)
                 }
             }
         }
     }
 }
-
-private const val KOVER_MERGED_REPORT_TASK_NAME = "koverMergedReport"
 
 internal val HubdleEnableableExtension.hubdleCoverage: HubdleConfigCoverageExtension
     get() = getHubdleExtension()